Make tests python version agnostic, ignore case
Some tests fail because they are python2 and 'python' might be a symlink
to python3. One idea is to make them executable and run them with ./. I
am not sure whether there is a way to do that without making them
executable. Some others fail because the case is not the same as the
reference file. Passing -i to diff should fix that.